media-video/ffmpeg: add 4.4.5-r2 based on 6.1.2-r1's ebuild
Similar bare losing a few options, albeit given this version is for
compat and normal users do not need the full set of features (they
should use ffmpeg-6+), skip a few inconvenient IUSE/options:

- chromium: needs a different patch/logic it seems
- glslang: ffmpeg-4 did not have shaderc support, meaning that this
  is the only alternative, but this breaks too easily
- mmal: was formerly kept due to lack of IUSE=soc, but would rather
  not keep the raspberrypi-userland dependency forever
- npp: not tested, and doubt anyone would want this on anything but
  latest ffmpeg (needs nvidia-cuda-toolkit)
- qsv: intelmedia-sdk is dead, and this version did not support libvpl
- amf: likely broken in this old version wrt bug #910271, assuming
  newer 4.4.5 did not fix it, but even if it did still little reason
  to support things like qsv/amf/nvenc as they are being left behind
  in old versions
- soc: plus all its logic, never had a patch for this version anyway
- vmaf: seems broken with current libvmaf version wrt bug #931137,
  and old libvmaf has some simd-related issues building (we *could*
  patch ffmpeg but at this point I feel little reason to for ffmpeg-4).
- --enable-avresample (bug #722274): off by default, and gone
  in later versions, no need to keep it enabled still
- --enable-lto: old version did not take =arg and, either way, keeping
  it on this older codebase with a lot of asm just seem like trouble

More would likely fit that list, but not trying to overdo until
something actually causes a problem.

Would've wanted to drop USE=nvenc due to its dependency on old
nv-codec-headers but rather mask it not to have to update revdeps
that depend on nvenc over this.

Hopefully not overlooked anything relevant, spent less time on
this than ffmpeg-7 and "downgrading" an ebuild is awkward.

Ultimately this version is going to be short lived, there is almost
nothing left (that we package) that requires it.

Closes: https://bugs.gentoo.org/722274
Closes: https://bugs.gentoo.org/910271
Closes: https://bugs.gentoo.org/931137
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>