media-video/ffmpeg-compat: new package, add 4.4.5, 6.1.2
Meant to be used with ffmpeg-compat.eclass, see its description.

4.4.5 likely won't see much use at this point if the ffmpeg-5
tracker is not missing anything, but adding anyway for now.

Note that this is not intended to provide libraries for old prebuilt
binaries, it is for building against using the eclass. If prebuilts
want to use this then they will likely need patchelf.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>