sys-kernel/scx: manually install openrc init.d script
Upstream's build system checks for openrc as a dependency before
installing the init.d script. This is not actually necessary, since
installing it on a non-openrc system is perfectly fine. Previously, scx
had an openrc use flag that toggled the openrc feature in meson. This
was not ideal since this is not a standardized use flag and openrc users
would have to manually enable it. Instead, disable the openrc meson
feature and install the init.d script manually in src_install.

Signed-off-by: Violet Purcell <vimproved@inventati.org>
Signed-off-by: Sam James <sam@gentoo.org>