media-libs/nv-codec-headers: add 12.2.72.0, 13.0.19.0
Two versions to support the range of (not-masked) in-tree
nvidia-drivers and what older ffmpeg versions may require to work
(albeit unneededed right now, ffmpeg-6.1.1 is broken with 13.0
but that version is due for cleanup either way).

Add a pkg_postinst to inform of driver support, but do not reintroduce
a nvidia-drivers RDEPEND (leads to complications and this is supposed
to be headers-only, runtime is optional).

Also simplify the ebuild, the makefile's `all` target does nothing
but a sed (not worth a src_compile), and the unintuitive LIBDIR=share
installs to /usr/<libdir>/pkgconfig like the `mv` does.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>