profiles: updates for media-video/ffmpeg IUSE changes
Adds IUSE="npp openmpt rabbitmq", mask where not keyworded (npp
can't gain more keywords, and the others kind of niche and not
really worth it).

soc mask is unneeded given 9999 no longer has the IUSE.

svg (rust) is due to no longer building with old librsvg, it *could*
be patched but this USE is not essential for the average user and
isn't worth maintaining.

There are a lot of pending cleanups due to removed IUSE, but that
has to wait until old versions removal.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>