media-video/ffmpeg: define USE=gnutls in metadata.xml
Global description does not fit with the logic used here.

USE=openssl's description was also updated in a prior commit
to state that it disables gnutls (aka, it is now mentioned both
ways).

I considered inverting the logic, but felt that it would be worse
unless USE=openssl is the default (i.e. you go out of your way
to enable non-default USE=openssl but it does nothing as long as
gnutls is set). And then, I am not inclined to change the default
given issues like bug #905113 had.

If this continues to confuse users we could always give it up
and do a REQUIRED_USE, albeit it may annoy some users with
conflicts.

Closes: https://bugs.gentoo.org/905113
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>