media-tv/kodi: use ffmpeg-compat
Does seem broken with ffmpeg-7 at a glance (was not sure if the
pinning was only to match the bundled version or because it was
actually broken) so let's use ffmpeg-compat instead to let users
upgrade.

kodi-9999 seems to pin to ffmpeg-7 instead, but not going to look
at making that work in the older version esp. since I do not use
this.

Keeping old revision which can be stabled without pulling
ffmpeg-compat:6 for stable users, not that it'd be a big deal.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>