games-emulation/ppsspp: use ffmpeg-compat
It is possible fixes already exist for ffmpeg-7 for this, but
not planning to look into that myself.

Arguably instead using bundled on ppsspp would not be a bad idea
given playing PSP's videos properly is flaky and ppsspp bundles
ffmpeg3 intentionally even though it could use newer. Then again
that's assuming the bundled copy even builds fine.

That aside, given no code changes here, plans are to sync
keywords back with stable when ready.

Bug: https://bugs.gentoo.org/948816
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>