media-video/ffmpeg: drop USE=doc if slotted
This USE is a no-op in ffmpeg-compat because of --disable-doc,
may as well hide it.

fwiw also hide the perl dependency while here given this is only
for man pages (pod2man) that are likewise disabled by that option.

Also drop man pages comment, left over from back when was still
installing man pages in testing -- but installing them there felt
weird so disabled them instead.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>