llvm-runtimes/libatomic-stub: new package, add 0
Clang(LLVM) implements atomic functions in compiler-rt, while GCC
provides a dedicated libatomic for it.
Some apps such as nodejs erroneously depend on GCC through libatomic.
LLVM's atomic builtins and libatomic are source-compatible so such
packages should be also buildable with LLVM.
It would be hard to fix all those packages, so let's deal with it
on compiler-rt side by installing a stub libatomic.a.

Bug: https://bugs.gentoo.org/911340
Signed-off-by: mojyack <mojyack@gmail.com>