dev-qt/qttools: fix clang selection for >=6.9
Qt added new logic for clang selection where it iterates through
a list of supported version and always picks the most recent one
from that list ignoring the eclass.

May be a better way, but for now just remove the version check.

Does not seem to have been backported to 6.8, so leaving that
one alone.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>