sci-visualization/xd3d: Fix build failures with GCC-15
By slapping -std=gnu17 on it. And fortran option to remove underscore
insertion in function names.
Patched relevant files to remove function name defines that fixed name
incompatibility (due to underscores) but also caused implicit declaration
errors.
Re-wrote makefiles - they were written like the author was lacking tab
key and worked around that. Now it finally stops on error, and not when
next job fails with bad dependencies.
Found and changed homepage to new URL.
Was unable to remove `which` - how do you do that in C shell??? It's not
POSIX, lacks `command` build-in.

Closes: https://bugs.gentoo.org/944746
Closes: https://bugs.gentoo.org/886609
Closes: https://bugs.gentoo.org/875236
Signed-off-by: NHOrus <jy6x2b32pie9@yahoo.com>