dev-qt/qt5compat: add IUSE=+gui
Often not bothering with USE=gui automagic, but this one was requested
in bug #951442 and qt5compat can be needed for a variety of non-gui
things too.

Using [gui=] rather than cmake_disable for two reasons, first because
the latter often has unexpected effects with Qt (side-disabling things
it shouldn't, tend to be fine with Qml but with Gui it can be messy)
and then so that gui revdeps do not have to think about whether USE=gui
is set on this package or not (if they depend on qtbase[gui] then it'll
be set) given unclear what packages need graphicaleffects5 controlled
by this.

Note that if a user setting USE=-gui globally and then only setting it
on specific packages like qtbase is reading this and wondering why
portage is giving confusing output, please set it on qt5compat too.

Closes: https://bugs.gentoo.org/951442
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>