media-video/dvdstyler: use ffmpeg-compat
There seem to be patches going around to fix this with ffmpeg-7,
but not going to look into testing these myself.

The app-cdr/dvdisaster dependency seem to be a bigger problem than
this though. I haven't looked too closely but it seems broken with
modern compilers and still depends on gtk+:2.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>