app-emulation/wine-*: bring back -mno-avx for PE files
Still see people have problems with this and believe wine passing
-mpreferred-stack-boundary=2 has only mitigated this. Or at least
with wine-10+ (but add it to 9 anyway as a precaution).

Do not believe it's worth the potential minor optimizations in
mingw-gcc built code, so let's just keep -mno-avx.

fwiw only revbumping the latest versions and stable, seem to be
less of an issue than before still and no need to rebuild all
slots.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>