media-video/dvdstyler: fix ffmpeg-compat detection
...when normal media-video/ffmpeg is not installed.

At a glance this does not use the `ffmpeg` command (FFMPEG_PATH
is not used in the sources that I can see, nor the resulting
binary), but ./configure will abort if it is not found. Thanks
to that, we can get away with passing SYSROOT and not revbumping
which would otherwise be incorrect for a runtime-used path.

Reconsidered just fixing with ffmpeg-7, but given this was previously
broken with "older" ffmpeg5 and patches are accumulating, this may
need special attention and runtime testing.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>