net-libs/gnutls: fix linkage (the lack of) for USE="zlib zstd brotli"
gnutls upstream supports these --with-* flags using the values
"yes/link/dlopen/no", and defaults to dlopen. This means that we
technically do not need the deps at runtime, like a kind of horrifying
"advertisement" for IUSE_RUNTIME in combination with build-only
dependencies. The actual goal here, I suppose, is that on Debian you can
unmerge zlib if you don't need gnutls support for it, and it's only a
"suggested" dependency, so a single binary build is useful for all
possible users. Gentoo has USE flags for this, and would have a hard
runtime dependency on the compression library either way, depending on
how it's built -- rebuilding without that USE is trivial.

So, for our use cases, it's much better to specify that we want to link
directly to the compression libs, as it makes iwdevtools happier, plays
nicely with preserved-libs, and avoids the frankly horrifying
possibility that some other library *uses* gnutls, and also uses zlib /
zstd / brotli, maybe via dlopen too, maybe dlclosing it multiple times.
If a plugin system was desirable, it would use gnutls plugins tbh (and
dlopen those, which in turn have genuine shlib linkage to the external
libs they provide a wrapper for).

So, using =link is all around safer with no downsides, let's do it.

Per Sam's advice: bumping via git mv.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>