sys-apps/xdg-desktop-portal: avoid gstreamer harder for now
gstreamer is searched for unconditionally, but we don't have gstreamer-pbutils
packaged yet, so that's a problem. Make it conditional for now (as it ought
to be upstream either way if they're going to have an option for it).

Noticed the upstream bug link and got the patch (tweaked style slightly
after) from flint2's PR at https://github.com/gentoo/gentoo/pull/41133.

Closes: https://bugs.gentoo.org/951609
Closes: https://bugs.gentoo.org/951611
Signed-off-by: Sam James <sam@gentoo.org>