profiles: downgrade amd64 LLVM profiles to 'dev'
These are experimental in reality, no real expectation the depgraph
be stable for them all the time, and the 'stable' marker confused someone
yet again on gentoo-user. No downsides to making it 'dev' and some upside
(less confusion for the odd user, though they shouldn't conflate the two
stability notions.. and also for developers to make clear they don't need
to really worry that much about it).

Signed-off-by: Sam James <sam@gentoo.org>