dev-qt/qttools: add comment about GUI applications handling
On one hand, the majority of users do not need the GUI assistant or
linguist, but on the other the cmake build system does not easily
split the GUI and their support libraries (Help/UiTools) very well
which needs us to build both.

As a non-optimal solution, *could* consider something like USE=gui
that would `rm` them but it feels kind of silly when already built
them and it would not really save any runtime-only dependencies.

Just keep a note about this situation for now.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>