sys-kernel/asahi-sources: add 6.13.7_p4
The added patches in files/ are needed to resolve conflicts between
genpatches and the Asahi patchset.
1740_x86-insn-decoder-test-allow-longer-symbol-names.patch is included
in the Asahi tree so simply revert it.
2980_kbuild-gcc15-gnu23-to-gnu11-fix.patch introduces a context conflict
in the top level Makefile. Revert it first so the Asahi changes can be
applied and then reapply it from 2981-kbuild-gcc15-gnu23-to-gnu11-fix.patch
which is refreshed against the Asahi tree.
Switch the Asahi Linux kernel patchset to '.diff' github URLs since the
the Asahi tree has duplicate rust commits to ease change management. The
duplicate commits are unfortunately reproduced in URLs like
https://github.com/AsahiLinux/linux/compare/v6.13.7...asahi-6.13.7-4.patch.
The '.diff' ones produce a simple diff between the two tags.

Signed-off-by: Janne Grunau <j@jannau.net>