virtual/package-manager: remove pkgcore as option
For long time, pkgcore stopped being maintained as a package manager.
Upstream (me) maintains pkgcore as a ebuild library in python. There are
known minor bugs, bad support for binary packages, and newer EAPIs might
not work perfectly like portage. The risk of users assuming pkgcore is
ready for them as package manager, and breaking something, is quite
high, so I prefer to drop pkgcore as an option here.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>