distutils-r1.eclass: Invert DISTUTILS_USE_PEP517 - backend mapping
Invert the eclass logic to map DISTUTILS_USE_PEP517 values to backends
rather than the other way around.  This is in preparation for making
it possible to override the backend used.  One side effect of this
is that in case of DISTUTILS_USE_PEP517 mismatch, we no longer provide
the "correct" value, and only indicate the backend used upstream.

Signed-off-by: Michał Górny <mgorny@gentoo.org>