dev-python/meson-python: fix erroneous build dep
cython is only needed by the testsuite. Futhermore, cython is treated as
a shell tool (it is looked up using meson's compiler language handling),
so it never made sense to require PYTHON_USEDEP.

The PYTHON_USEDEP forces recompiling cython if you want to install
meson-python for multiple impls. It also means if you try for a pure
py3.13 system (which I'm trying to do for the binhost) but need a small
handful of 3.12 packages you get conflicts, because meson-python holds
back having a pure 3.13 experience for cython.

Fixes: 34bf7b6af8ca51d45b8696665c0823511d067211
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>