app-emulation/wine-*: update live
IUSE=bluetooth usage is header-only to support creating rfcomm
sockets, and is automagic when bluetooth.h/rfcomm.h are detected
without passing cache variables.

Technically does not need the REQUIRED_USE on dbus, but bluetooth
is also used through dbus and that makes it clearer to users that
they shouldn't disable dbus for bluetooth support (not depending
on bluez at runtime given essentially a optfeature, and actual
bluetooth users wouldn't need us to pull the dependency for them).

Also sync comment with wine-vanilla that randomly spotted.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>