dev-tex/biblatex: add media-fonts/corefonts as USE 'doc' conditional dep
The biblatex documentation at doc/latex/biblatex/biblatex.tex uses
"Courier New" as mono font. Therefore, we add media-libs/corefonts as
conditional BDEPEND.

We also mark two bugs now closed, that are related to biblatex.pdf not
being installed if USE=doc. Technically, this was fixed with the when
biblatex-3.20-r1 was first added but was forgotten back then. Hence we
do it now.

Closes: https://bugs.gentoo.org/938927
Closes: https://bugs.gentoo.org/945334
Signed-off-by: Florian Schmaus <flow@gentoo.org>