app-shells/fish: sync live ebuild, fix link failure with gcc
This happens when we rebuild the code with Fish's CMake wrapper (we need
that to install auxilary files and run test), since cargo.eclass and
Fish's build system put the target executables in different output
directories.

First we patch CMake so that it doesn't even try to execute cargo for
it's targets and just retain the target to copy the executables over into
the CMake build directory.

Also we need to set the cargo output directory in the CMake scripts to
point to the target directory of cargo.eclass so it reuses the already
build binaries.

And since Fish sometimes uses a custom cargo build profile which alters
the target directory in a way not handled by cargo.eclass we remove that
as well.

Signed-off-by: idealseal <realidealseal@protonmail.com>
Closes: https://github.com/gentoo/gentoo/pull/41068
Signed-off-by: Sam James <sam@gentoo.org>