media-video/mpv: simplify ebuild a bit
_multi is redundant with what the meson.build already does,
may as well drop it entirely (e.g. if feature egl and drm
are enabled, it enables egl-drm and doesn't really do anything
else special, not much sense in passing egl-drm ourselves).

This *shouldn't* change anything, so no need for a revbump.
If it does, then it's an oversight that needs fixing.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>