net-im/qtox: update for ffmpeg usedep changes
Previously ffmpeg had both A) USE=libv4l and B) USE=v4l. A is for using
media-libs/libv4l, and B is for direct (more limited) support only with
linux-headers.

With the new layout, B is now always enabled (no USE) assuming linux,
and A was renamed to be enabled through USE=v4l (matching other
packages that do not make this distinction).

Then the new USE=v4l (and old libv4l) is currently masked on musl, and
this causes problems here (would require masking the package on musl).
But then B is always enabled (which is what it used to depend on), so
we just need to remove it.

It'd be more correct to depend on the exact revisions that changed the
meaning of USE=v4l, but feel it's not worth the extra complications
(we can't use e.g. v4l(+) due to the change in meaning).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>