dev-vcs/git: add copies-harder patch
I wasn't planning on adding this to dev-vcs/git, but after polling
a few times on IRC in #gentoo-dev, it seems people want it.

It was rejected upstream (after several revisions; we hope it'll be
accepted in future) because they don't see the value in having a config
option for this, but when working in a repo like ::gentoo, it's natural
to want it as a default.

It seems okay to patch it in downstream unconditionally given you have
to actively opt-in to *using* it by adding the relevant config option -
this was also the opinion of others which helped persuade me to do this.

The docs added by the patch also mention that it's patched in and not
upstream.

Details are within the patch itself, but quoting some of it:
> diff: implement config.diff.renames=copies-harder
>
> This patch adds a config value for 'diff.renames' called 'copies-harder'
> which make it so '-C -C' is in effect always passed for 'git log -p',
> 'git diff', etc.
>
> This allows specifying that 'git log -p', 'git diff', etc should always act
> as if '-C --find-copies-harder' was passed.

(I also ended up having to manually distribute rebased versions of this
patch on new releases, so it's better all-round to do this, I think).

Signed-off-by: Sam James <sam@gentoo.org>