media-sound/frescobaldi: add 4.0.0
Hardly know this so not really tested beyond checking if it starts
and seems normal, please report if find any downstream packaging
issues, or major issues that we should backport fixes for before
potential stable (after reporting upstream if not already).

Please also CC me for downstream bugs for now, haven't added myself as
maintainer given may not handle this package long term (just trying to
get it in a good state to be easier to handle for other maintainers).

If no major issues, hoping to stabilize before the py3.13 target switch
in May (3.13 is not on 3.3.0, untested) and soon follow that by 3.3.0
removal to allow pyqtwebengine (Qt5) cleanup.

Closes: https://bugs.gentoo.org/905637
Closes: https://bugs.gentoo.org/926681
Closes: https://bugs.gentoo.org/943686
Closes: https://bugs.gentoo.org/952546
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>