media-video/ffmpeg: slightly clarify S comment
This is to allow copying the ebuild as-is to ffmpeg-compat and
not have to maintain them as two different ebuilds. If same
version, the only difference can at most be keywords if in middle
of stabilization but otherwise these should match as well.

ffmpeg-compat ebuilds could potentially be simplified when the
equivalent ffmpeg version is dropped though.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>