net-misc/curl: add ECH and HTTPSRR USE flags; prepare for 8.13.0
HTTPS Resource Records are a type of DNS record that delivers configuration
information and parameters for how to access a service via HTTPS.
This should make QUIC in particular faster to establish connections due to
being able to retrieve the h3 endpoint without bouncing off an alt-svc first.

These may be fetched via DNS over HTTPS or via the cURL ADNS resolver
(c-ares); the threaded resolver is disabled in Gentoo if HTTP RR is
enabled as a request will need to be made via c-ares to fetch the
records anyway.

Encrypted Client Hello is now available via Rustls. ECH is a protocol extension
to Transport Layer Security (TLS) which encrypts part of the handshake and masks
the Server Name Indication (SNI) that is used to negotiate a TLS session.

ECH info may be served via HTTPS RR or DoH.

This commit also refactors the TLS configure option handling in the ebuild
to avoid passing `--without` entries for enabled backends.

Additionally USE=`sasl-scram` is added, using libgsasl (net-misc/gsasl)
to extend SASL SCRAM-SHA support, and USE=`sslv3` is dropped.

Signed-off-by: Matt Jolly <kangie@gentoo.org>