dev-qt/qtwebengine: add 6.9.0
Also including backport for QTBUG-135047 (x11 pixmap leak) which
is fixed in 6.9.9999 but hasn't been included in the 6.9.0 release.

Not bothering putting it in the patchset given would need to make
new one for 6.9.9999 already.

The glibc-2.41 patch is also still not included, but it should
whenever 6.9.9999 gets a qtwebengine-chromium submodule update.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>