dev-qt/qttools: fix clang detection with libcxx
Happened to notice this while testing 6.9.0 with llvm-musl, and
this been broken since 6.8.3 (with USE=clang) that had the same
change picked to.

CMake Error at cmake/FindWrapLibClang.cmake:72 (find_package):
  find_package called with invalid argument "20.1.1+libcxx"

At first I thought this was caused by our sed, but that specific
line is unmodified and is using ${LLVM_VERSION} which has the
+libcxx bit appended (or at least, it does with llvm:20). This
is Qt's attempt at getting a "matching" version for Clang.

Given we are trying to remove version specifications to let the
eclass pick them instead (will always match), may as well just
remove it as well.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>