sci-electronics/spice: correctly error when compiling fails
The makefiles for this project are from 1989, back when software
developers still knew how to write real build systems. For example, real
build systems are intelligent enough to:

- mark loads of rules as "-command", which tells make to not fail if the
  command fails
- run make by concatenating makefile fragments and piping it to `make -f -`
- recursively run more makes via concatenation
- pass -k to all make invocations, so that we try to build as much as
  possible even if we can't succeed in the end
- mandatorily require running make from a shell script that sets up
  everything, somehow

There is only one slight problem with this otherwise brilliant design.
Namely, none of it is a good idea.

ANYWAY, this software does not build due to Modern C violations. It is
also entirely uninteresting to invest serious time in the build system
for this thing, or even attempt to report it, since the software exists
as a wikipedia "historically significant events" page plus random
third-party download sites and web.archive.org and a progression of
modernized and extended forks. The actual project was declared finished
and done with in 1993.

I'm told keeping the package around is important because, as the father
of an entire industry, it's important to have as a point of reference
whenever doing various kinds of work. Still, the darned thing could do a
bit better about actually compiling. Step 1 is to error out on compile
failures, rather than succeeding and then failing to run the built
executables / have the input files for "dobin" and erroring out with
inscrutable messages.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>