sci-electronics/spice: "fix" Modern C issues by declaring defeat
The software is extremely crusty. As mentioned in the previous commit,
it hasn't been maintained since 1993 and there are derivative projects
which likely solve these issues. The code has "worked" since 1993 and I
suspect unlikely to ever be of much interest to new platforms, given
that it's only even useful as a point of reference compared to its
modern derivatives. Keep in mind the code is so old that it's from back
when the most recent brand new C revision supported implicit function
declarations.

Trying to solve its C issues seems unworthwhile. Just -std=c89 it.
Actually, gnu89 since it depends on `#define linux` to detect the
platform. :P

Closes: https://bugs.gentoo.org/919301
Closes: https://bugs.gentoo.org/944870
Closes: https://bugs.gentoo.org/945068
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>