net-analyzer/wireshark: revert to old zlib due to conflict with minizip-ng
Discussion of bug #953390 has revealed that Wireshark's minizip-ng
feature currently relies on it being built with minizip compatibility,
which is unfortunately not easily possible at the moment.
Technically this could be solved by linking both zlib[minizip] and
zlib-ng, but down that path lies madness, so just consistently revert
to zlib/minizip.

Closes: https://bugs.gentoo.org/953390
Signed-off-by: Holger Hoffstätte <holger@applied-asynchrony.com>