media-libs/libsdl2: add 2.32.54 (aka sdl2-compat, unkeyworded)
For testing, looking at the upstream issue tracker it is probably
still early to have it replace libsdl2 entirely (there's also the
qemu static-libs issue, see ebuild comment) but should work in most
cases nonetheless.

Feel free to opt-in and ideally report issues to sdl's upstream
rather than Gentoo unless it looks like a packaging issue (which
may well exist, not tested much yet).

wrt versions, to clarify, >=x.y.50 are sdl2-compat while <x.y.50
is normal libsdl2. There could be a 2.34.0 and 2.34.50 or so in
the future, so compat is not necessarily *all* >=2.32.50.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>