media-libs/libsdl2: add pulse/pipewire IUSE to sdl2-compat for now
Missed that a package depends on libsdl2[pipewire], albeit it sounds
questionable. Ideally would want to remove all the audio IUSE.

pulse is not requested by anything, but it could confuse users if it
is missing when there is other audio USE. Not going as far as adding
things like oss or sndio (note that they are usable regardless).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>