media-libs/libsdl3: drop IUSE=static-libs
qemu still using this doesn't really make sense and, if it migrates
to libsdl3, the seemingly legacy usage should be cleaned up.

Unlikely that it works right too given nothing checks that all the
other libraries it needs to link with are available statically.

Can always be re-added later if we really need it, but for now do not
want to encourage anything to depend on this.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>