profiles: add asahi VIDEO_CARDS value and add arch use masks
This will be used to build support for Apple AGX GPUs, as found
in Apple Silicon Macs and supported by the Asahi and Honeykrisp
Mesa drivers (OpenGL and Vulkan respectively).

The arch based used masks are a little questionable since the driver
will have virtgpu DRM native-context support. For now just unmask it on
the native arm64 architecture and x86 and amd64 for emulation of x86
software.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>
Signed-off-by: Janne Grunau <j@jannau.net>
Signed-off-by: Matt Turner <mattst88@gentoo.org>