dev-libs/boost: wire up USE=test/test-full
This adds the initial bits to run boost's test suites,
all of which are different. We filter out failing tests
and make everything work with the same options as the regular
build, which required some creative hackery to run legacy
tests that were missing a main().

Closes: https://bugs.gentoo.org/953468
Signed-off-by: Holger Hoffstätte <holger@applied-asynchrony.com>
Part-of: https://github.com/gentoo/gentoo/pull/41616
Closes: https://github.com/gentoo/gentoo/pull/41616
Signed-off-by: Sam James <sam@gentoo.org>