dev-libs/libffi: don't add executable stack on HPPA
When building libffi on HPPA with `-Wl,--warn-warn-execstack`, we get:
```
ld: warning: src/pa/.libs/linux.o: missing .note.GNU-stack section implies executable stack
ld: NOTE: This behaviour is deprecated and will be removed in a future version of the linker
``

This is because pa/linux.S lacks a .GNU-stack note so it gets one added
automatically that states an executable stack is needed.

That becomes more problematic with glibc-2.41 which forbids dlopen()
of a library with an executable stack, and libffi is commonly dlopen()'d,
especially by Python.

I suspect the reason it didn't show up on Debian is that since February,
Debian has been building Binutils with --disable-default-execstack.

Closes: https://bugs.gentoo.org/953805
Signed-off-by: Sam James <sam@gentoo.org>