1/2: rename USE=sound to oss
"sound" is a _very_ generic flag name with an obvious description.
The Open Sound System flag still preserves the information about
the fact that this will enable sound but also holds the fact that
it is done using a specific software.

Also doing this avoids confusion since USE=alsa will also provide
sound functionality. This makes the user more aware about the
underlying technology making it easier to made an educated guess
or a conscious decision.