games-board/scid: enable py3.13
py3.13 is only used by optional scripts with no tests (the package
has tests, but they are unrelated to python), at a glance they do
not appear broken but hard to say (not properly tested).

...except pgnfix which may possibly(?) be broken with both 3.12 and
3.13, would need to recheck with 3.11 and test with an actual pgn
input file. That script also need a patch to work with python3 at
all, arguably we should probably just skip it at this point.

Also, for some reason 5.0.2 had re-gained py3.9 despite it wasn't in
4.7.0, so dropping that as well.

Closes: https://bugs.gentoo.org/952479
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>