games-puzzle/brainworkshop: add 5.0.3_p20250317, enable py3.13
While testing for py3.13, noticed that this was broken entirely by
>=pyglet-2 regardless. Then it needs multiple backports for them
to apply and things to work properly (the piglet-2 commit is not
enough). Simpler to do a snapshot for now, may still have issue but
given 5.0.3 is broken entirely and <piglet-2 is gone from Gentoo
it doesn't make anything worse.

Closes: https://bugs.gentoo.org/952492
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>