x11-drivers/nvidia-drivers: add 575.51.02 (unkeyworded beta)
Beta, and unless there is a major reason to, we do not keyword these
given all the warnings NVIDIA gives about the potential issues they
*could* have (incl. damaging hardware).

Feel free to opt-in if wanted.

Note that this ships with a egl-wayland+egl-x11 snapshot, meaning our
release versions are older. Have not verified what interaction this
may have (will check more closely when keywording a non-beta).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>