dev-python/lit: enable py3.13 for old versions
Tests all pass the same regardless of version. 15 and 16 do have warning
messages fixed in 17 e.g.:
```
SyntaxWarning: invalid escape sequence '\s'
```

but these aren't an issue for 3.13 support and there's no specific
timeline for when this will be converted to an error anyway. Hopefully
whenever they decide to make this a fatal error, the final remaining
desperate stragglers on llvm 15 will have been ported, but we'll see.

Acked-by: Sam James <sam@gentoo.org>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>