toolchain.eclass: extend JIT location hack to libgdiagnostics too
libgccjit is strange in that it's a non-internal library installed
by GCC and isn't intended to be interacted with via some magic options
like -fopenmp or whatever, but just linked like a regular library.

libgdiagnostics is the same, so extend the same treatment to it.

Thanks to tdr for testing.

Closes: https://bugs.gentoo.org/953823
Signed-off-by: Sam James <sam@gentoo.org>