sys-auth/elogind: enable py3.13 in old version
The latest is still masked on musl profiles, so need to do this
one as well. Seems fine with 3.13 at a glance, not that going to
test much for this old version.

Revbump is due to python-any-r1 PYTHON_USEDEP deps and needing
to ensure that users' VDB is updated to avoid conflicts.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>